refactor(cleanup): renamed some files for clarity, minor fixes
